requirements.txt
amqp==5.3.1
asgiref==3.5.2
billiard==4.2.4
black==22.6.0
celery==5.5.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==4.1.3
django-celery-beat==2.8.1
django-crontab==0.7.1
django-extensions==3.2.1
django-timezone-field==7.2.1
djangorestframework==3.14.0
idna==3.4
kombu==5.5.4
mypy-extensions==0.4.3
packaging==26.0
pathspec==0.10.1
platformdirs==2.5.3
prompt_toolkit==3.0.52
pyngrok==7.5.0
python-crontab==3.3.0
python-dateutil==2.8.2
python-dotenv==1.2.1
pytz==2022.6
PyYAML==6.0.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.12
vine==5.1.0
wcwidth==0.5.2
redis==6.2.0
